Move processResources up to MessageService. Tweak tests to match new expected format.
